Create README.md
                 VEHICLE SECURITY APP
This project is chosen with regard to the increasing cases of stolen Bicycles in NITT. Due to inadequate security neither can we save them from getting stolen nor can we find out the missing cycle. The project will be boon to the NITT students as it will help them to save the time, efforts and money, to find the stolen cycle.
  My project will contain two software’s- first one to generate a unique id for the vehicle, the student has purchased and the other software will be present with the security personnel’s at the main gate to validate that the cycle which is been taken in or out of the campus is not a stolen one. The main gate personnel can then accordingly allow the vehicle to cross the gates.
This project will be prepared using the C++ concepts using different classes for both the software. The first software will be to register the vehicle and will generate a unique vehicle id and also a card which will have a unique number. At the gates the second software will be used, where the student will be asked the card and the card number when entered will generate the unique vehicle id which was given to the student and was imprinted on the cycle. This will validate whether the cycle belongs to that person or not.